requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.0
mediapipe>=0.8.9
numpy>=1.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
ffmpeg-python>=0.2.0
torch>=1.9.0
torchaudio>=0.9.0
whisper>=1.0.0 